Bug 1299823 - Rename PR_X* symbols, remove unused ones. r=kaie
None of these symbols are in a public header, they are not NSPR_API, and
only available on XP_UNIX builds, which in the vast majority of cases,
means built with -fvisibility=hidden (i.e. the symbols are not even
exported in the first place).

The convention for non-exported symbols is to start with _PR rather than
PR, so we rename those symbols.

The PR_XWait, PR_XNotify and PR_XNotifyAll symbols are however never used,
so we just remove them.

Differential Revision: https://phabricator.services.mozilla.com/D132448